+++ src/components.ts
@@ -32,6 +32,8 @@ import { createThirdPartyProvidersGraphFetcherComponent } from './adapters/third
 import { createThirdPartyProvidersServiceFetcherComponent } from './adapters/third-party-providers-service-fetcher'
 import { createThirdPartyProvidersStorage } from './logic/third-party-providers-storage'
 import { createIdentityComponent } from './adapters/identity'
+import { createProfilesComponent } from './adapters/profiles'
+import { createHasherComponent } from './adapters/hasher'
 
 // Initialize all the components of the app
 export async function initComponents(
@@ -124,6 +126,20 @@ export async function initComponents(
   )
 
   const identity = createIdentityComponent()
+  const profiles = await createProfilesComponent({
+    metrics,
+    content,
+    theGraph,
+    config,
+    fetch,
+    ownershipCaches,
+    thirdPartyProvidersStorage,
+    logs,
+    wearablesFetcher,
+    emotesFetcher
+  })
+
+  const hasher = createHasherComponent()
   return {
     config,
     logs,
@@ -155,6 +171,8 @@ export async function initComponents(
     catalystsFetcher,
     poisFetcher,
     nameDenylistFetcher,
-    identity
+    identity,
+    profiles,
+    hasher
   }
 }
